Restore game conversation view wiring
